Refactor code organization into modular structure
- Move Color class to src/radiometry/Color.h
- Move Camera class to src/radiometry/Camera.h
- Move PPMWriter class to src/io/PPMWriter.h
- Update rt.cpp to use new modular includes
- Create INTERFACE libraries for header-only modules
- Set up proper dependency chain: io -> radiometry -> geometry
- All tests pass and rt app works correctly
