+++ src/apps/CMakeLists.txt
@@ -9,7 +9,7 @@ target_link_libraries(test_app PRIVATE glfw)
 
 # Raytracer app
 add_executable(rt rt.cpp)
-target_link_libraries(rt PRIVATE geometry)
+target_link_libraries(rt PRIVATE geometry radiometry io)
 target_compile_features(rt PRIVATE cxx_std_23)
 
 # Metal Triangle App - build separately with: swiftc -o metal_triangle metal_triangle.swift -framework Cocoa -framework Metal -framework MetalKit
